Alot of refactoring
