--- src/main/java/se/accelerateit/signup6/model/EmailMessage.java
@@ -1,14 +1,0 @@
-package se.accelerateit.signup6.model;
-
-import lombok.Data;
-
-import java.net.URL;
-
-@Data
-public class EmailMessage {
-
-    private String to;
-    private String subject;
-    private String message;
-    private URL eventURL;
-}
