+++ src/main/java/se/accelerateit/signup6/scheduling/ScheduledEvents.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import se.accelerateit.signup6.dao.EventMapper;
 import se.accelerateit.signup6.email.emailUtil.UserFilter;
-import se.accelerateit.signup6.email.impl.EmailSenderServiceImpl;
+import se.accelerateit.signup6.email.EmailSenderService;
 import se.accelerateit.signup6.model.Event;
 import se.accelerateit.signup6.model.User;
 
@@ -16,11 +16,11 @@ import java.util.List;
 public class ScheduledEvents {
 
     private final EventMapper eventMapper;
-    private final EmailSenderServiceImpl senderService;
+    private final EmailSenderService senderService;
     private final UserFilter userFilter;
 
     @Autowired
-    public ScheduledEvents(EventMapper eventMapper, UserFilter userFilter, EmailSenderServiceImpl senderService) {
+    public ScheduledEvents(EventMapper eventMapper, UserFilter userFilter, EmailSenderService senderService) {
         this.eventMapper = eventMapper;
         this.senderService = senderService;
         this.userFilter = userFilter;
